Fix ppc64 build warnings
Fix these ppc64 32 bit build warnings:

 kexec/arch/ppc64/kexec-zImage-ppc64.c: In function 'zImage_ppc64_load':
 kexec/arch/ppc64/kexec-zImage-ppc64.c:164: warning: cast to pointer from integer of different size
 kexec/arch/ppc64/kexec-elf-ppc64.c: In function 'elf_ppc64_load':
 kexec/arch/ppc64/kexec-elf-ppc64.c:121: warning: integer constant is too large for 'unsigned long' type
 kexec/arch/ppc64/kexec-elf-ppc64.c:237: warning: cast from pointer to integer of different size
 kexec/arch/ppc64/kexec-elf-ppc64.c:276: warning: cast from pointer to integer of different size
 kexec/arch/ppc64/kexec-elf-ppc64.c:283: warning: cast from pointer to integer of different size
 kexec/arch/ppc64/kexec-elf-ppc64.c:287: warning: cast from pointer to integer of different size
 kexec/arch/ppc64/kexec-elf-ppc64.c:341: warning: format '%lx' expects type 'long unsigned int', but argument 3 has type 'uint64_t'
 kexec/arch/ppc64/kexec-elf-ppc64.c:352: warning: format '%ld' expects type 'long int', but argument 5 has type 'size_t'
 kexec/arch/ppc64/crashdump-ppc64.c:45: warning: integer constant is too large for 'long' type
 kexec/arch/ppc64/crashdump-ppc64.c:46: warning: integer constant is too large for 'long' type
 kexec/arch/ppc64/crashdump-ppc64.c:56: warning: integer constant is too large for 'long' type
 kexec/arch/ppc64/crashdump-ppc64.c:57: warning: integer constant is too large for 'long' type

Tested on PS3 (ppc64) with 32 and 64 bit builds.

+++ kexec/arch/ppc64/kexec-zImage-ppc64.c
@@ -161,7 +161,7 @@ int zImage_ppc64_load(FILE *file, int argc, char **argv, void **ret_entry,
 	segment->mem = (void *) load_loc;
 	segment->memsz = memsize;
 	segment->bufsz = filesize;
-	*ret_entry = (void *)((uint64_t)elf.e_entry);
+	*ret_entry = (void *)(uintptr_t)elf.e_entry;
 	*ret_nr_segments = i - 1;
 	free(ph);
 	return 0;
